+++ packages/Skeleton/src/types.ts
@@ -1,6 +1,11 @@
+import { ThemeType } from '@blaze-react/themes';
+
 export interface SkeletonProps {
   classes: string;
   children?: JSX.Element | JSX.Element[];
+  variant?: string | 'default';
+  theme?: ThemeType;
+  overrides?: string[];
 }
 
 export interface InputProps {