+++ packages/themes/src/context/themeProvider.tsx
@@ -1,14 +1,18 @@
 import { ThemeType } from '../types';
-import React, { useState } from 'react';
+import React from 'react';
 import { ThemeContext } from './themeContext';
 
 interface ThemeProviderProps {
   theme: ThemeType;
+  showSkeleton: boolean;
   children: unknown;
+  setTheme?: (theme: ThemeType) => void;
 }
 
-export function ThemeProvider({ theme, children }: ThemeProviderProps): JSX.Element {
-  const [currentTheme, setTheme] = useState(theme);
-
-  return <ThemeContext.Provider value={{ theme: currentTheme, setTheme: setTheme }}>{children}</ThemeContext.Provider>;
+export function ThemeProvider({ theme, setTheme, showSkeleton, children }: ThemeProviderProps): JSX.Element {
+  return (
+    <ThemeContext.Provider value={{ theme: theme, setTheme: setTheme, showSkeleton: showSkeleton }}>
+      {children}
+    </ThemeContext.Provider>
+  );
 }