+++ packages/themes/src/theme/components/skeleton.ts
@@ -1,0 +1,7 @@
+export const skeleton = {
+  variants: {
+    default: {
+      container: ['bg-gray-400', 'w-full', 'h-full', 'animate-pulse'],
+    },
+  },
+};