+++ packages/Skeleton/CHANGELOG.md
@@ -3,6 +3,129 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.8.0-alpha.49](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.48...v0.8.0-alpha.49) (2021-10-08)
+
+**Note:** Version bump only for package @blaze-react/skeleton
+
+
+
+
+
+# [0.8.0-alpha.48](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.47...v0.8.0-alpha.48) (2021-10-08)
+
+**Note:** Version bump only for package @blaze-react/skeleton
+
+
+
+
+
+# [0.8.0-alpha.47](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.46...v0.8.0-alpha.47) (2021-10-07)
+
+**Note:** Version bump only for package @blaze-react/skeleton
+
+
+
+
+
+# [0.8.0-alpha.46](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.45...v0.8.0-alpha.46) (2021-10-06)
+
+**Note:** Version bump only for package @blaze-react/skeleton
+
+
+
+
+
+
+
+**Note:** Version bump only for package @blaze-react/skeleton
+
+
+
+
+
+# [0.8.0-alpha.45](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.44...v0.8.0-alpha.45) (2021-10-06)
+
+**Note:** Version bump only for package @blaze-react/skeleton
+
+
+
+
+
+# [0.8.0-alpha.44](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.43...v0.8.0-alpha.44) (2021-10-06)
+
+**Note:** Version bump only for package @blaze-react/skeleton
+
+
+
+
+
+# [0.8.0-alpha.43](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.42...v0.8.0-alpha.43) (2021-10-06)
+
+**Note:** Version bump only for package @blaze-react/skeleton
+
+
+
+
+
+# [0.8.0-alpha.42](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.41...v0.8.0-alpha.42) (2021-10-06)
+
+**Note:** Version bump only for package @blaze-react/skeleton
+
+
+
+
+
+# [0.8.0-alpha.41](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.39...v0.8.0-alpha.41) (2021-10-06)
+
+**Note:** Version bump only for package @blaze-react/skeleton
+
+
+
+
+
+# [0.8.0-alpha.40](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.39...v0.8.0-alpha.40) (2021-10-06)
+
+**Note:** Version bump only for package @blaze-react/skeleton
+
+
+
+
+
+# [0.8.0-alpha.39](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.38...v0.8.0-alpha.39) (2021-10-05)
+
+**Note:** Version bump only for package @blaze-react/skeleton
+
+
+
+
+
+# [0.8.0-alpha.38](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.37...v0.8.0-alpha.38) (2021-10-05)
+
+**Note:** Version bump only for package @blaze-react/skeleton
+
+
+
+
+
+
+
+**Note:** Version bump only for package @blaze-react/skeleton
+
+
+
+
+
+# [0.8.0-alpha.37](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.36...v0.8.0-alpha.37) (2021-10-05)
+
+
+### Features
+
+* add showSkeleton to ThemeProvider ([5c611ed](https://github.com/thebyte9/blaze-components-react/commit/5c611ed9b4232ab5358cf1c4f7b8120d27ced163))
+
+
+
+
+
 # [0.8.0-alpha.36](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.35...v0.8.0-alpha.36) (2021-09-08)
 
 **Note:** Version bump only for package @blaze-react/skeleton